pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-wtf>=1.2.2",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "flask-login>=0.6.3",
    "python-telegram-bot==20.8",
    "stripe>=11.6.0",
    "pyjwt>=2.10.1",
    "flask-security>=5.6.0",
    "telegram>=0.0.1",
    "sqlalchemy>=2.0.38",
]
